ingestion/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pytz
google-cloud-bigquery==3.23.1
google-cloud-bigquery-storage==2.25.0
